requirements_test.txt
# Requirements for MQTT testing scripts
paho-mqtt>=1.6.0


